requirements.txt
dnspython>=2.2.1
colorama>=0.4.6
pyOpenSSL>=23.1.1
cryptography>=41.0.3
requests>=2.31.0
requests-toolbelt>=0.10.1
